@@ -34,7 +34,7 @@
     "ansi-colors": "4.1.1",
     "autoprefixer": "9.7.3",
     "babel-eslint": "10.0.3",
-    "babel-plugin-istanbul": "5.2.0",
+    "babel-plugin-istanbul": "6.0.0",
     "babelify": "10.0.0",
     "bluebird": "3.7.2",
     "body-parser": "1.19.0",
@@ -46,11 +46,12 @@
     "cookie-parser": "1.4.4",
     "cssnano": "4.1.10",
     "del": "5.1.0",
-    "eslint": "6.7.2",
+    "eslint": "6.8.0",
     "eslint-plugin-google-camelcase": "0.0.2",
     "eslint-plugin-prettier": "3.1.2",
-    "eslint-plugin-swg-internal": "file:./build-system/eslint-rules",
     "eslint-plugin-sort-imports-es6-autofix": "0.5.0",
+    "eslint-plugin-sort-requires": "2.1.0",
+    "eslint-plugin-swg-internal": "file:./build-system/eslint-rules",
     "express": "4.17.1",
     "fancy-log": "1.3.3",
     "fetch-mock": "7.7.3",
@@ -64,7 +65,7 @@
     "gulp-git": "2.10.0",
     "gulp-help": "chmontgomery/gulp-help#gulp4",
     "gulp-if": "3.0.0",
-    "gulp-load-plugins": "2.0.1",
+    "gulp-load-plugins": "2.0.2",
     "gulp-nightwatch": "1.2.0",
     "gulp-rename": "2.0.0",
     "gulp-replace": "1.0.0",
@@ -96,20 +97,20 @@
     "karma-super-dots-reporter": "0.2.0",
     "lazypipe": "1.0.2",
     "minimist": "1.2.0",
-    "mocha": "6.2.2",
+    "mocha": "7.0.0",
     "morgan": "1.9.1",
     "nodemon": "2.0.2",
     "plugin-error": "1.0.1",
-    "postcss": "7.0.25",
+    "postcss": "7.0.26",
     "postcss-import": "12.0.1",
     "prettier": "1.19.1",
     "request": "2.88.0",
-    "rollup": "1.27.13",
+    "rollup": "1.28.0",
     "rollup-plugin-commonjs": "10.1.0",
     "rollup-plugin-node-resolve": "5.2.0",
     "shuffle-seed": "1.1.6",
-    "sinon": "7.5.0",
-    "sinon-chai": "3.3.0",
+    "sinon": "8.0.4",
+    "sinon-chai": "3.4.0",
     "through2": "3.0.1",
     "touch": "3.1.0",
     "vinyl-buffer": "1.0.1",
