@@ -27,8 +27,8 @@ const log = require('fancy-log');
 const path = require('path');
 const watch = require('gulp-watch');
 const {gitDiffNameOnlyMaster} = require('../git');
-const {isTravisBuild} = require('../travis');
 const {green, yellow, cyan, red} = require('ansi-colors');
+const {isTravisBuild} = require('../travis');
 
 const isWatching = argv.watch || argv.w || false;
 const options = {
@@ -233,6 +233,7 @@ lint.description = 'Validates against Google Closure Linter';
 lint.flags = {
   'watch': '  Watches for changes in files, validates against the linter',
   'fix': '  Fixes simple lint errors (spacing etc)',
+  'files': '  Lints just the specified files',
   'local_changes': '  Lints just the files changed in the local branch',
   'quiet': '  Suppress warnings from outputting',
 };
