@@ -331,11 +331,12 @@ describes.realWin('installRuntime legacy', {}, env => {
 describes.realWin('Runtime', {}, env => {
   let win;
   let runtime;
-  const loggedEvents = [];
+  let loggedEvents = [];
 
   beforeEach(() => {
     win = env.win;
     runtime = new Runtime(win);
+    loggedEvents = [];
     sandbox
       .stub(ClientEventManager.prototype, 'logEvent')
       .callsFake(event => loggedEvents.push(event));
@@ -471,11 +472,25 @@ describes.realWin('Runtime', {}, env => {
         active: null,
         data: null,
       });
-      expect(loggedEvents.length).to.equal(1);
-      expect(loggedEvents[0]).to.deep.equal({
+      expect(loggedEvents.length).to.equal(2);
+      const payEvent =
+        loggedEvents[0].eventType === AnalyticsEvent.IMPRESSION_PAYWALL
+          ? loggedEvents[0]
+          : loggedEvents[1];
+      const startEvent =
+        loggedEvents[0].eventType === AnalyticsEvent.IMPRESSION_PAGE_LOAD
+          ? loggedEvents[0]
+          : loggedEvents[1];
+      expect(payEvent).to.deep.equal({
         eventType: AnalyticsEvent.IMPRESSION_PAYWALL,
         eventOriginator: EventOriginator.PUBLISHER_CLIENT,
         isFromUserAction: null,
+        additionalParameters: null,
+      });
+      expect(startEvent).to.deep.equal({
+        eventType: AnalyticsEvent.IMPRESSION_PAGE_LOAD,
+        eventOriginator: EventOriginator.SWG_CLIENT,
+        isFromUserAction: false,
         additionalParameters: null,
       });
       expect(logger).to.be.instanceOf(Logger);
@@ -938,7 +953,7 @@ describes.realWin('ConfiguredRuntime', {}, env => {
       try {
         await configPromise;
       } catch (e) {}
-      expect(eventCount).to.equal(1);
+      expect(eventCount).to.equal(2);
     });
 
     it('should not log when config rejected', async () => {
@@ -1256,7 +1271,7 @@ describes.realWin('ConfiguredRuntime', {}, env => {
 
       it('work for 1 entitlement', async () => {
         entitlements = [
-          new Entitlement('', ['product1'], '{"productId":"token1"}'),
+          new Entitlement('google', ['product1'], '{"productId":"token1"}'),
         ];
         analyticsMock
           .expects('setSku')
@@ -1266,9 +1281,9 @@ describes.realWin('ConfiguredRuntime', {}, env => {
 
       it('work for multiple entitlement', async () => {
         entitlements = [
-          new Entitlement('', ['product1'], '{"productId":"token1"}'),
-          new Entitlement('', ['product2'], '{"productId":"token2"}'),
-          new Entitlement('', ['product3'], '{"productId":"token3"}'),
+          new Entitlement('google', ['product1'], '{"productId":"token1"}'),
+          new Entitlement('google', ['product2'], '{"productId":"token2"}'),
+          new Entitlement('google', ['product3'], '{"productId":"token3"}'),
         ];
         analyticsMock
           .expects('setSku')
@@ -1283,6 +1298,14 @@ describes.realWin('ConfiguredRuntime', {}, env => {
 
       it('kind of work for non-JSON entitlement', async () => {
         entitlements = [new Entitlement('', ['product1'], 'token1')];
+        analyticsMock
+          .expects('setSku')
+          .withExactArgs('unknown subscriptionToken')
+          .once();
+      });
+
+      it('missing product ID in SwG entitlement', async () => {
+        entitlements = [new Entitlement('google', ['product1'], 'token1')];
         analyticsMock
           .expects('setSku')
           .withExactArgs('unknown subscriptionToken')
