@@ -14,71 +14,14 @@
  * limitations under the License.
  */
 
-/* @const */
-const toString_ = Object.prototype.toString;
-
-/**
- * Returns the ECMA [[Class]] of a value
- * @param {*} value
- * @return {string}
- */
-function toString(value) {
-  return toString_.call(value);
-}
-
-/**
- * Determines if value is actually an Array.
- * @param {*} value
- * @return {boolean}
- */
-export function isArray(value) {
-  return Array.isArray(value);
-}
-
-/**
- * Converts an array-like object to an array.
- * @param {?IArrayLike<T>|string} arrayLike
- * @return {!Array<T>}
- * @template T
- */
-export function toArray(arrayLike) {
-  if (!arrayLike) {
-    return [];
-  }
-  const array = new Array(arrayLike.length);
-  for (let i = 0; i < arrayLike.length; i++) {
-    array[i] = arrayLike[i];
-  }
-  return array;
-}
-
 /**
  * Determines if value is actually an Object.
  * @param {*} value
  * @return {boolean}
  */
 export function isObject(value) {
-  return toString(value) === '[object Object]';
-}
-
-/**
- * Determines if value is of number type and finite.
- * NaN and Infinity are not considered a finite number.
- * String numbers are not considered numbers.
- * @param {*} value
- * @return {boolean}
- */
-export function isFiniteNumber(value) {
-  return typeof value === 'number' && isFinite(value);
-}
-
-/**
- * Determines if value is of FormData type.
- * @param {*} value
- * @return {boolean}
- */
-export function isFormData(value) {
-  return toString(value) === '[object FormData]';
+  const str = Object.prototype.toString.call(value);
+  return str === '[object Object]';
 }
 
 /**
@@ -101,15 +44,17 @@ export function isEnumValue(enumObj, s) {
 /**
  * True if the value is a function.
  * @param {*} value
+ * @return {boolean}
  */
 export function isFunction(value) {
-  return value !== null && typeof value === 'function';
+  return typeof value === 'function';
 }
 
 /**
  * True if the value is either true or false.
  * @param {?*} value
+ * @return {boolean}
  */
 export function isBoolean(value) {
-  return value === true || value === false;
+  return typeof value === 'boolean';
 }
