@@ -25,21 +25,10 @@ const commands = {
       '16 Top Spots for Hiking - The Scenic'
     );
   },
-  viewOffers: function() {
-    return this.pause(1000)
-      .log('Viewing offers')
-      .switchToFrame('[src*="about:blank"]', 'SwG outer iFrame')
-      .switchToFrame('[src*="offersiframe"]', 'SwG inner iFrame');
-  },
   selectOffer: function() {
     return this.viewOffers()
       .log('Selecting "Basic Access" offer')
-      .pause(10000)
-      .waitForElementPresent({
-        selector: '.qLPyoc',
-        time: 10000,
-        poll: 100,
-      })
+      .waitForElementPresent('.qLPyoc')
       .click('.qLPyoc')
       .pause(1000);
   },
