@@ -29,8 +29,7 @@ module.exports = {
       .viewFirstArticle()
       .waitForElementPresent('@swgIFrame', 'Found SwG iFrame')
       .viewOffers()
-      .pause(5000)
-      .waitForElementPresent('.K2Fgzb', 10000, 'Found offer carousel')
+      .waitForElementPresent('.K2Fgzb', 'Found offer carousel')
       .assert.containsText('.K2Fgzb', 'Subscribe with your Google Account')
       .assert.containsText('.wlhaj.I3RyHc', 'Already subscribed?')
       .assert.containsText('.amekj', 'Basic Access')
